+++ src/BabylonCpp/include/babylon/bones/skeleton.h
@@ -11,9 +11,9 @@
 namespace BABYLON {
 
 class BABYLON_SHARED_EXPORT Skeleton
-  : public ::std::enable_shared_from_this<Skeleton>,
-    public IAnimatable,
-    public IDisposable {
+    : public ::std::enable_shared_from_this<Skeleton>,
+      public IAnimatable,
+      public IDisposable {
 
 public:
   Skeleton(const string_t& name, const string_t& id, Scene* scene);
@@ -40,7 +40,7 @@ public:
    */
   int getBoneIndexByName(const string_t& name);
 
-  void createAnimationRange(const string_t& name, int from, int to);
+  void createAnimationRange(const string_t& name, float from, float to);
   void deleteAnimationRange(const string_t& name, bool deleteFrames = true);
   AnimationRange* getAnimationRange(const string_t& name);
